requirements.txt
requests>=2.31.0
Pillow>=10.0.0
python-docx>=1.1.0
pywin32>=306
